refactor: move notification builders to dedicated notifications folder
