+++ webtrit_callkeep_android/android/src/main/kotlin/com/webtrit/callkeep/api/background/ProxyBackgroundCallkeepApi.kt
@@ -8,7 +8,7 @@ import com.webtrit.callkeep.PDelegateBackgroundServiceFlutterApi
 import com.webtrit.callkeep.common.ActivityHolder
 import com.webtrit.callkeep.common.helpers.Platform
 import com.webtrit.callkeep.common.models.CallMetadata
-import com.webtrit.callkeep.services.NotificationService
+import com.webtrit.callkeep.notifications.NotificationService
 import com.webtrit.callkeep.services.AudioService
 
 /**
