+++ webtrit_callkeep_android/android/src/main/kotlin/com/webtrit/callkeep/connection/PhoneConnection.kt
@@ -17,7 +17,7 @@ import com.webtrit.callkeep.api.foreground.TelephonyForegroundCallkeepApi
 import com.webtrit.callkeep.common.helpers.Platform
 import com.webtrit.callkeep.common.models.CallMetadata
 import com.webtrit.callkeep.services.AudioService
-import com.webtrit.callkeep.services.NotificationService
+import com.webtrit.callkeep.notifications.NotificationService
 
 /**
  * Represents a phone connection for handling telephony calls.
