+++ webtrit_callkeep_android/android/src/main/kotlin/com/webtrit/callkeep/notifications/ActiveCallNotificationBuilder.kt
@@ -1,4 +1,4 @@
-package com.webtrit.callkeep.services.notification
+package com.webtrit.callkeep.notifications
 
 import android.app.Notification
 import android.app.NotificationChannel
